+++ src/TurboSmsChannel.php
@@ -1,45 +1,27 @@
 <?php
 
+declare(strict_types=1);
+
 namespace NotificationChannels\TurboSms;
 
 use Illuminate\Events\Dispatcher;
 use Illuminate\Notifications\Events\NotificationFailed;
-use NotificationChannels\TurboSms\Exceptions\CouldNotSendNotification;
 use Illuminate\Notifications\Notification;
+use NotificationChannels\TurboSms\Exceptions\CouldNotSendNotification;
 
 class TurboSmsChannel
 {
-    /**
-     * @var TurboSmsApi
-     */
-    protected $smsApi;
+    public function __construct(
+        protected TurboSmsApi $smsApi,
+        protected Dispatcher $events,
+    ) {}
 
-    /**
-     * @var Dispatcher
-     */
-    protected $events;
-
-    /**
-     * TurboSmsChannel constructor.
-     * @param TurboSmsApi $smsApi
-     * @param Dispatcher $events
-     */
-    public function __construct(TurboSmsApi $smsApi, Dispatcher $events)
-    {
-        $this->smsApi = $smsApi;
-        $this->events = $events;
-    }
     /**
      * Send the given notification.
      *
-     * @param  mixed  $notifiable
-     * @param  Notification  $notification
-     *
      * @throws CouldNotSendNotification
-     *
-     * @return array|null
      */
-    public function send($notifiable, Notification $notification)
+    public function send(mixed $notifiable, Notification $notification): ?array
     {
         try {
             $recipient = $this->getRecipient($notifiable);
@@ -48,27 +30,33 @@ class TurboSmsChannel
             if (is_string($message)) {
                 $message = new TurboSmsMessage($message);
             }
+
             if (! $message instanceof TurboSmsMessage) {
                 throw CouldNotSendNotification::invalidMessageObject($message);
             }
 
             return $this->smsApi->sendMessage($recipient, $message);
 
-        } catch (Exception $exception) {
-            $event = new NotificationFailed($notifiable, $notification, 'TurboSms', ['message' => $exception->getMessage(), 'exception' => $exception]);
+        } catch (\Throwable $exception) {
+            $event = new NotificationFailed(
+                $notifiable,
+                $notification,
+                'TurboSms',
+                ['message' => $exception->getMessage(), 'exception' => $exception]
+            );
 
-            if (function_exists('event')) { // Use event helper when possible to add Lumen support
-                event($event);
-            } else {
-                $this->events->fire($event);
-            }
+            $this->events->dispatch($event);
         }
+
+        return null;
     }
 
-    protected function getRecipient($notifiable)
+    protected function getRecipient(mixed $notifiable): string
     {
-        if ($notifiable->routeNotificationFor('TurboSms')) {
-            return $notifiable->routeNotificationFor('TurboSms');
+        $phone = $notifiable->routeNotificationFor('TurboSms', null);
+
+        if ($phone) {
+            return $phone;
         }
 
         if (isset($notifiable->phone)) {